fixes for ai_pipeline tests
